src/main: Version 4.0, Final adjustements.
    + Some minor corrections have been made. At worst case, if
      the code does not function properly with TimerOne.h, the branch
      will be rolled back to about 3.4 version.
    - I was unable to make AD interrupt work properly with TimerOne.h.

+++ src/main/Utils.h
@@ -30,5 +30,5 @@
 #define KEY_a               16
 
 float Voltage(uint8_t PIN) { 
-    return (float)analogRead(PIN) * (float)ref_voltage / (float)Max_ADC_Ret;
+    return ( analogRead(PIN) * ref_voltage ) / (float)Max_ADC_Ret;
 }
